Move figures into separate file
