+++ pages/blog/[id].js
@@ -1,6 +1,5 @@
 import fs from 'fs'
 import matter from 'gray-matter'
-import dynamic from 'next/dynamic'
 import { MDXRemote } from 'next-mdx-remote'
 import { serialize } from 'next-mdx-remote/serialize'
 import path from 'path'
@@ -18,6 +17,7 @@ import {
 import { Code, Pre } from '@carbonplan/prism'
 
 import { postMetadata, POSTS_PATH } from '../../utils/mdx-utils'
+import figures from '../../figures'
 
 const COMPONENTS = {
   code: Code,
@@ -28,39 +28,6 @@ const COMPONENTS = {
   Figure,
   FigureCaption,
   TableCaption,
-}
-
-const figures = {
-  'climate-trace-release': {
-    Map: dynamic(() => import('../../posts/climate-trace-release/figure')),
-  },
-  'open-lidar-biomass': {
-    Chart: dynamic(() => import('../../posts/open-lidar-biomass/figure')),
-  },
-  'maps-library-release': {
-    MapDemo2d: dynamic(() =>
-      import('../../posts/maps-library-release/maps-demo-2d')
-    ),
-    MapDemo4d: dynamic(() =>
-      import('../../posts/maps-library-release/maps-demo-4d')
-    ),
-  },
-  'soil-protocols-added': {
-    RecommendationTable: dynamic(() =>
-      import('../../posts/soil-protocols-added/recommendation-table')
-    ),
-    ScoreSummary: dynamic(() =>
-      import('../../posts/soil-protocols-added/score-summary')
-    ),
-  },
-  'ton-year-ncx': {
-    TableHundred: dynamic(() =>
-      import('../../posts/ton-year-ncx/table-hundred')
-    ),
-    TableThousand: dynamic(() =>
-      import('../../posts/ton-year-ncx/table-thousand')
-    ),
-  },
 }
 
 const PostPage = ({ id, source, frontMatter, number }) => {
